change name + organize code
